# fitness-proxy/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Пример сервера для документации",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://apache.org"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/proxy",
    "paths": {
        "/management/cache": {
            "delete": {
                "description": "Удаляет все записи из кеша в памяти",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cache-Management"
                ],
                "summary": "Очистить кеш",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/management/cache_setting/{id}": {
            "get": {
                "description": "Возвращает настройки кеша для указанного ObjectID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cache-Management"
                ],
                "summary": "Получить настройки кеша по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectID настройки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/management/cache_settings/purge": {
            "delete": {
                "description": "Удаляет весь кеш для совпадений по началу пути (например, /api/users. Удалит в том числе /api/users/uuid)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cache-Management"
                ],
                "summary": "Удалить кеш по началу пути",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Путь для удаления настроек",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/management/cache_settings/{id}": {
            "put": {
                "description": "Изменяет ttl у опредленной настройки кеша по id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cache-Management"
                ],
                "summary": "Изменить TTL настройки кеша по id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID настройки кеша (Hex)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое значение TTL в JSON",
                        "name": "ttl_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет настройку кеша для указанного ObjectID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cache-Management"
                ],
                "summary": "Удалить настройку кеша по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectID настройки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/management/clients": {
            "get": {
                "description": "Возвращает список самых активных IP-адресов с сортировкой по количеству запросов и логами блокировок",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Получить топ активных клиентов и нарушителей",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/service.ClientStats"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/management/ip_access/verify-captcha": {
            "post": {
                "description": "Проверяет запрос от фронтенда и выставляет сессионную куку доверия для IP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security"
                ],
                "summary": "Подтвердить прохождение капчи",
                "parameters": [
                    {
                        "description": "Данные верификации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CaptchaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/management/logs": {
            "get": {
                "description": "Возвращает последние 100 структурированных JSON-логов из MongoDB с возможностью фильтрации по уровню и IP",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Получить асинхронные логи аудита",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по уровню логов (DEBUG, INFO, WARN, ERROR)",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по IP-адресу клиента",
                        "name": "ip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.AccessLog"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "error: Не удалось прочитать логи из БД",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/management/metrics": {
            "get": {
                "description": "Возвращает историю RPS, задержки и трафика за последние 60 секунд для графиков",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitoring"
                ],
                "summary": "Получить динамические метрики",
                "responses": {}
            }
        },
        "/management/reload": {
            "get": {
                "description": "Обновляет актуальные правила IP",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-Management"
                ],
                "summary": "Подгрузить правила IP",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/management/rules": {
            "get": {
                "description": "Возвращает список всех правил IP",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-Management"
                ],
                "summary": "Получить все правила IP",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.IPRule"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет новый IP в черный, белый или серый список",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP-Management"
                ],
                "summary": "Добавить правило IP",
                "parameters": [
                    {
                        "description": "Данные правила",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.IPRule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/management/rules/{id}": {
            "delete": {
                "description": "Удаляет запись из MongoDB по её ObjectID",
                "tags": [
                    "IP-Management"
                ],
                "summary": "Удалить правило IP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ObjectID правила",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.CaptchaRequest": {
            "type": "object",
            "required": [
                "ip"
            ],
            "properties": {
                "ip": {
                    "type": "string"
                }
            }
        },
        "model.AccessLog": {
            "type": "object",
            "properties": {
                "id": {},
                "ip": {
                    "type": "string"
                },
                "latency": {
                    "type": "integer"
                },
                "level": {
                    "description": "Например: \"info\", \"warning\", \"error\"",
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "reason": {
                    "description": "Например: \"Blacklisted by administrator\"",
                    "type": "string"
                },
                "requestID": {
                    "description": "Уникальный ID для отслеживания запроса",
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.IPRule": {
            "type": "object",
            "properties": {
                "burst": {
                    "description": "максимальный всплеск",
                    "type": "integer"
                },
                "comment": {
                    "description": "Для чего это правило (необязательно, но полезно)",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "network": {
                    "description": "Например: \"192.168.1.0/24\"",
                    "type": "string"
                },
                "rate": {
                    "description": "количество запросов в секунду",
                    "type": "number"
                },
                "type": {
                    "description": "\"white\", \"black\", \"grey\"",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "service.ClientStats": {
            "type": "object",
            "properties": {
                "blocked_blacklist": {
                    "description": "Забанен по IP",
                    "type": "integer"
                },
                "blocked_rate_limit": {
                    "description": "Превысил лимит RPS/RPM",
                    "type": "integer"
                },
                "blocked_requests": {
                    "type": "integer"
                },
                "bytes_transferred": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "total_requests": {
                    "type": "integer"
                }
            }
        }
    }
}